Add upload img

+++ app/Http/Controllers/PostsController.php
@@ -3,27 +3,35 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
-       // with DB library
-       //  $posts = DB::select('SELECT * fROM posts')
-
-         // return
+        // with DB library
+        // $posts = DB::select('SELECT * fROM posts')
+        // return
         // $posts = Post::orderBy('title', 'desc')->take(1)->get();
         // $posts = Post::orderBy('title', 'desc')->get();
         // $posts = Post::where('title', 'Post two')->get();
 
-        //pagination
+        // pagination
         $posts = Post::orderBy('created_at', 'desc')->paginate(2);
-
-
-     //$posts = Post::all();
+       // $posts = Post::all();
         return view('posts.index')->with('posts', $posts);
     }
 
@@ -37,12 +45,38 @@ class PostsController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999',
         ]);
+
+        // Handle file upload
+        if($request->hasFile('cover_image'))
+        {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }else{
+            $fileNameToStore = 'noimage.jpg';
+        }
+
 
         // Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -51,6 +85,11 @@ class PostsController extends Controller
     public function show($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.show')->with('post', $post);
     }
 
@@ -67,11 +106,33 @@ class PostsController extends Controller
             'body' => 'required',
         ]);
 
+        // Handle file upload
+        if($request->hasFile('cover_image'))
+        {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
         // Create Post
         $post = Post::find($id);
-
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image'))
+        {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -80,6 +141,18 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if(auth()->user()->id !==$post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        // Delete file
+        if($post->cover_image != 'noimage.jpg')
+        {
+            // Delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
